(PC-20410) refacto(adaptOffersPlaylistParameters): move adaptOffersPlaylistParameters closer to where it is used
