+++ src/libs/algolia/fetchAlgolia/fetchMultipleOffers/adaptOffersPlaylistParameters.ts
