+++ src/libs/algolia/fetchAlgolia/fetchMultipleOffers/useAdaptOffersPlaylistParameters.ts
@@ -3,8 +3,8 @@ import { useCallback } from 'react'
 import { OffersModuleParameters } from 'features/home/types'
 import { useMaxPrice } from 'features/search/helpers/useMaxPrice/useMaxPrice'
 import { SearchState } from 'features/search/types'
+import { adaptOffersPlaylistParameters } from 'libs/algolia/fetchAlgolia/fetchMultipleOffers/adaptOffersPlaylistParameters'
 import { useGeolocation } from 'libs/geolocation'
-import { adaptOffersPlaylistParameters } from 'libs/search/adaptOffersPlaylistParameters'
 import { useGenreTypeMapping, useSubcategoryLabelMapping } from 'libs/subcategories/mappings'
 
 export const useAdaptOffersPlaylistParameters = () => {
